+++ Modules/Utils/Sources/PocketCastsUtils/Feature Flags/FeatureFlag.swift
@@ -323,11 +323,7 @@ public enum FeatureFlag: String, CaseIterable {
         case .libroFm:
             false
         case .grdb:
-            #if DEBUG
             true
-            #else
-            false
-            #endif
         case .encourageAccountCreation:
             true
         case .notificationsRevamp:
@@ -391,8 +387,6 @@ public enum FeatureFlag: String, CaseIterable {
             "use_podcast_html_description"
         case .podcastViewChanges:
             "podcast_view_changes_2025"
-        case .grdb:
-            "grdb_testflight"
         default:
             rawValue.lowerSnakeCased()
         }
@@ -406,9 +400,9 @@ extension FeatureFlag: OverrideableFlag {
 
     public var canOverride: Bool {
         switch self {
-            // GRDB can only change in TestFlight versions
+            // GRDB can only change to `false` in non-TestFlight versions
             case .grdb:
-                Self.isTestFlight
+                !Self.isTestFlight
             default:
                 true
         }