+++ Modules/Utils/Sources/PocketCastsUtils/Feature Flags/FeatureFlag.swift
@@ -363,7 +363,7 @@ public enum FeatureFlag: String, CaseIterable {
         case .retryWithoutUserAgent:
             true
         case .userSatisfactionSurvey:
-            false
+            true
         }
     }