+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+7.95
+-----
+
+
 7.94
 -----
 - Retry failed requests without user agent [#3323](https://github.com/Automattic/pocket-casts-ios/pull/3323)